Django/Back-End/siftit/requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
mysql-connector==2.1.6
mysqlclient==1.3.12
psycopg2==2.7.5
python3-utils==0.4.0
requests==2.18.4
ujson==1.35
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
gunicorn
dj_database_url
python-decouple
django-toolbelt